Override PATH to avoid Debian-induced headaches
Debian adds RabbitMQ wrappers that only allow you to run RabbitMQ
commands as either root or the rabbitmq user. These not-exactly-useful
wrappers still run the command when another user executes them, albeit
without any arguments. For rabbitmq-server this is especially sinister
as that actually runs the server, but resulting in it mysteriously
ignoring the -detached flag that we were using.