Be creative with hostname
It turns out Travis workers do not use a qualified hostname, and
RabbitMQ gets way more confused than most people about this, especially
considering we never feed it anything but the short hostname anyhow.

By appending a .x to the hostname, RabbitMQ magically behaves, even
without adding this hostname to /etc/hosts or anything.

The suffix .x was chosen to be as short as possible, as changing the
hostname might fail with the very describe "hostname too long" error.