Fix a non-deterministic integration test
There was no guarantee that the messages would be evenly divided between the three buckets, but that was what the test was asserting. This version just checks that at least one bucket got at least one message – because the point of the test is just to show that custom demultiplexers can be used, not that the specific demultiplexer works.